iPhone: Fixed the answer submissions for the round so it successfully competes in a round. Also made the progress bar update with the time left to answer for points

+++ src/clients/iphone/distrivia/Classes/DistriviaAPI.h
@@ -19,7 +19,7 @@
 + (BOOL) registerWithData:(GameData*)gd user:(NSString*)userName pass:(NSString*)pass;
 + (BOOL) joinPublicWithData:(GameData*)gd;
 + (BOOL) statusWithData:(GameData*)gd;
-+ (BOOL) answerWithData:(GameData*)gd answer:(NSString*)answer timeTaken:(double)time;
++ (BOOL) answerWithData:(GameData*)gd answer:(NSString*)answer timeTaken:(int)time;
 
 + (NSMutableURLRequest*) createPost:(NSString*)post urlFrag:(NSString*)urlFragment;
